Added Macro for initializing callbacks
- The Macro [derive(init_callbacks)] parses the file Callbacks.toml

- Callback.toml contains descriptions of the individual callbacks in string format:
  - The associated variable
  - Condition for the execution of the callback
  - The callback function

- These strings are parsed using a parser generated with LALRPOP
  - The parser is still very rudimentary at this point, as it only needs to differentiate between variables (to be able to add the proper access functions) and everything else, which needs to be valid Rust code
  - The output of the parser is a new String which combines the modified variable names and the rest of the previous string
  - This string is then, as usual, parsed to a TokenStream and returned by the macro

- This feature compiles, but is as of now UNTESTED
