+++ aurora_hal/aurora_hal_macros/src/Callbacks.toml
@@ -1,0 +1,4 @@
+[1]
+var = "Process.Sensor1.pressure"
+condition = "Process.Sensor1.pressure > 5"
+callback = "Process.Sensor1.pressure = 5"
