+++ aurora_hal/aurora_hal_macros/src/CentralDataStruct.toml
